Fix notification spam after unmuting repos
When you unmute a repo it was showing all the historical activity as
new notifications, which was super annoying. Now tracking unmute
timestamps and using that as the checkDate when fetching activity.

Also cleaning up old unmute entries from storage so they don't pile up.
Tests added to make sure the cleanup works.